refactor: LottoConfig 클래스명 변경, InputValidator 패키지 이동
AppConfig
- LottoConfig -> AppConfig로 변경

InputValidator
- view -> common.validator로 패키지 이동
