+++ src/main/java/lotto/Application.java
@@ -1,10 +1,10 @@
 package lotto;
 
-import lotto.common.config.LottoConfig;
+import lotto.common.config.AppConfig;
 
 public class Application {
     public static void main(String[] args) {
-        LottoConfig config = new LottoConfig();
+        AppConfig config = new AppConfig();
         config.getLottoController().run();
     }
 }
