+++ src/test/java/lotto/view/InputValidatorTest.java
@@ -2,6 +2,7 @@ package lotto.view;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import lotto.common.validator.InputValidator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
